Removed the big machinePrinterKlippy... provider.
This also means that each card is now rendered by itself and on demand! This also shows the slivers for each card I worked on for the alst 4 months.
However the error handling for Printer/Klippy is now different. Klippy ASYNC-Errors are now handeled by the klippy_state_widget. For printer-Async_errors this is still WIP/TODO